Enhanced UI and some changes
