+++ package.json
@@ -27,6 +27,7 @@
     "react-toastify": "8.2.0",
     "redux": "4.1.2",
     "redux-persist": "6.0.0",
+    "sass": "^1.55.0",
     "web-vitals": "0.2.4",
     "workbox-background-sync": "5.1.4",
     "workbox-broadcast-update": "5.1.4",
