+++ src/components/ITI/ITIOptions/OptionCard/index.module.scss
@@ -1,0 +1,53 @@
+@mixin onMobile {
+    @media (max-width: 599px) {
+        @content;
+    }
+}
+
+.cardContainer {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;
+    padding: 2rem;
+    width: 18%;
+    border-radius: 1rem;
+    color: #00695c;
+    text-align: center;
+    font-weight: 600;
+    cursor: pointer;
+    border: 4px solid #00695c;
+    transition: all 0.3s ease-in-out;
+    box-shadow: rgb(38, 57, 77) 0px 20px 30px -10px;
+
+    @include onMobile() {
+        width: 100%;
+        flex-direction: row;
+        padding: 1rem;
+        margin-bottom: 2rem;
+        box-shadow: none;
+
+        & div:nth-child(2) {
+            width: 80%;
+            padding: 0rem 1rem;
+        }
+    }
+
+    svg {
+        height: 6rem;
+        width: 6rem;
+        margin-bottom: 2rem;
+
+        @include onMobile() {
+            height: 4rem;
+            width: 4rem;
+            margin: 0rem;
+        }
+    }
+
+    &:hover {
+        color: #FFF;
+        background: #00695c;
+    }
+
+}
