+++ src/components/ITI/ITIOptions/index.module.scss
@@ -1,0 +1,17 @@
+@mixin onMobile {
+    @media (max-width: 599px) {
+        @content;
+    }
+}
+
+.container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 0rem 10rem;
+
+    @include onMobile() {
+        flex-direction: column;
+        padding: 0rem 1rem;
+    }
+}
